refactor class
