+++ src/main/java/phanes/replay/gathering/controller/GatheringController.java
@@ -1,19 +1,16 @@
 package phanes.replay.gathering.controller;
 
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
-import phanes.replay.gathering.dto.request.GatheringCreateRq;
-import phanes.replay.gathering.dto.request.GatheringRq;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 import phanes.replay.gathering.dto.response.GatheringRs;
 import phanes.replay.gathering.service.GatheringService;
 
-import java.util.HashMap;
+import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Map;
 
 @RestController
 @RequiredArgsConstructor
@@ -21,72 +18,10 @@ import java.util.Map;
 public class GatheringController {
 
     private final GatheringService gatheringService;
-    @PostMapping
-    public ResponseEntity<?> creatGathering(
-            @RequestBody @Valid GatheringCreateRq request,
-            @RequestHeader(value = "Authorization", required = false) String authHeader,
-            BindingResult bindingResult
-            ){
-
-        System.out.println("컨트롤러 도달");
-        try {
-            if (bindingResult.hasErrors()) {
-                // 어떤 필드에서 에러 났는지 출력
-                return ResponseEntity.badRequest().body(bindingResult.getAllErrors());
-            }
-
-
-            Long userId = 1L;
-
-
-            gatheringService.createGathering(request,userId);
-
-            Map<String, String> response = new HashMap<>();
-            response.put("message", "모임 생성 성공!");
-
-            return ResponseEntity.ok(response);
-        } catch (Exception e) {
-            e.printStackTrace(); // 예외 메시지 확인 가능
-
-            Map<String, String> response = new HashMap<>();
-            response.put("error", "잘못된 요청");
-
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(response);
-        }
-
-    }
 
     @GetMapping
-    public ResponseEntity<?> getGatheringList(
-            @RequestParam(name = "sortBy", required = false) String sortBy
-        , @RequestParam(name = "keyword", required = false) String keyword
-        , @RequestParam(name = "location", required = false) String location
-        , @RequestParam(name = "date", required = false) String date
-        , @RequestParam(name = "genre", required = false) String genre
-        , @RequestParam(name = "limit", required = false, defaultValue = "10") Integer limit
-        , @RequestParam(name = "offset", required = false, defaultValue = "0") Integer offset){
-
-        try{
-
-            GatheringRq request = new GatheringRq();
-            request.setSortBy(sortBy);
-            request.setKeyword(keyword);
-            request.setLocation(location);
-            request.setDate(date);
-            request.setGenre(genre);
-            request.setLimit(limit);
-            request.setOffset(offset);
-
-            // 서비스
-
-            List<GatheringRs> response = gatheringService.getGatheringList(request);
-            return ResponseEntity.ok(response);
-        } catch (Exception e) {
-            e.printStackTrace();
-            Map<String, String> error = new HashMap<>();
-            error.put("error", "잘못된 요청");
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(error);
-        }
+    public List<GatheringRs> gatheringList(@AuthenticationPrincipal Long userId, @RequestParam String sortBy, @RequestParam String keyword, @RequestParam String city, @RequestParam String state, @RequestParam LocalDateTime date, @RequestParam String genre, @RequestParam Integer limit, @RequestParam Integer offset) {
+        userId = userId == null ? 0L : userId;
+        return gatheringService.getGatheringList(userId, sortBy, keyword, city, state, date, genre, limit, offset);
     }
-
 }
