+++ src/main/java/phanes/replay/gathering/persistence/repository/GatheringCommentRepository.java
@@ -1,4 +1,4 @@
-package phanes.replay.gathering.repository;
+package phanes.replay.gathering.persistence.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
