+++ src/main/java/phanes/replay/gathering/persistence/repository/GatheringContentRepository.java
@@ -1,4 +1,4 @@
-package phanes.replay.gathering.repository;
+package phanes.replay.gathering.persistence.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
